+++ apps/probability/law/uniform_law.h
@@ -18,10 +18,10 @@ public:
   I18n::Message parameterNameAtIndex(int index) override;
   I18n::Message parameterDefinitionAtIndex(int index) override;
   float evaluateAtAbscissa(float x) const override;
-  bool authorizedValueAtIndex(float x, int index) const override;
+  bool authorizedValueAtIndex(double x, int index) const override;
   void setParameterAtIndex(float f, int index) override;
-  float cumulativeDistributiveFunctionAtAbscissa(float x) const override;
-  float cumulativeDistributiveInverseForProbability(float * probability) override;
+  double cumulativeDistributiveFunctionAtAbscissa(double x) const override;
+  double cumulativeDistributiveInverseForProbability(double * probability) override;
 private:
   constexpr static float k_diracMaximum = 10.0f;
   constexpr static float k_diracWidth = 0.001f;
