+++ apps/shared/curve_view_cursor.cpp
@@ -9,22 +9,22 @@ CurveViewCursor::CurveViewCursor() :
 {
 }
 
-float CurveViewCursor::x() {
+double CurveViewCursor::x() {
   return m_x;
 }
 
-float CurveViewCursor::y() {
+double CurveViewCursor::y() {
   return m_y;
 }
 
-void CurveViewCursor::moveTo(float x, float y) {
+void CurveViewCursor::moveTo(double x, double y) {
   m_x = clipped(x, false);
   m_y = clipped(y, true);
 }
 
-float CurveViewCursor::clipped(float x, bool canBeInfinite) {
-  float maxValue = canBeInfinite ? INFINITY : k_maxFloat;
-  float clippedX = x > k_maxFloat ? maxValue : x;
+double CurveViewCursor::clipped(double x, bool canBeInfinite) {
+  double maxValue = canBeInfinite ? INFINITY : k_maxFloat;
+  double clippedX = x > k_maxFloat ? maxValue : x;
   clippedX = clippedX < - k_maxFloat ? -maxValue : clippedX;
   return clippedX;
 }
